requirements.txt

streamlit==1.32.0
langgraph==0.0.32
langchain-openai==0.1.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
pydantic==2.6.4
pandas==2.2.1
